pyproject.toml
[project]
name = "vsea-gtm-analyzer-ai"
version = "0.1.0"
description = "ADK-powered GTM analyzer agents for VentureSea (pitch-deck and URL-grounded modes)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "uvicorn[standard]>=0.34.0",
    "google-adk>=1.14.0",
    "google-genai>=1.0.0",
    "google-cloud-storage>=2.18.0",
    "python-pptx>=1.0.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.10.0",
    "python-json-logger>=3.3.0",
    "httpx>=0.28.0",
    "tomli>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "coverage>=7.6.0",
    "ruff>=0.6.0",
]

[tool.taskipy.tasks]
app = "uvicorn main:app --reload --port 8003"
test = "coverage run -m pytest tests"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
